requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
newspaper3k==0.2.8
numpy==1.19.5
openai==0.7.0
pandas==1.2.3
requests==2.25.1
soupsieve==2.2
urllib3==1.26.4
google-api-python-client==2.18.0